@@ -3,7 +3,7 @@ class Admin::InterfacesController < Admin::BaseController
   before_action :get_interface, only: [:update]
 
   def update
-    if @interface.update_attributes(interface_params)
+    if @interface.update(interface_params)
       flash[:notice] = "Position mise à jour avec succès"
       redirect_back(fallback_location: admin_root_path)
     else
@@ -20,5 +20,5 @@ class Admin::InterfacesController < Admin::BaseController
   def get_interface
     @interface = params[:classname].constantize.find params[:id]
   end
- 
+
 end
