@@ -30,7 +30,7 @@ class Admin::ProfilesController < Admin::BaseController
   end
 
   def update
-    if @profile.update_attributes(profile_params)
+    if @profile.update(profile_params)
       flash[:notice] = "Profil mis à jour avec succès"
       redirect_to params[:continue].present? ? edit_admin_profile_path(@profile) : admin_profiles_path
     else
@@ -43,7 +43,7 @@ class Admin::ProfilesController < Admin::BaseController
   end
 
   def update_configuration
-    if @profile.update_attributes(profile_params)
+    if @profile.update(profile_params)
       flash[:notice] = "Profil mis à jour avec succès"
       redirect_to params[:continue].present? ? edit_configuration_admin_profile_path(@profile) : admin_profiles_path
     else
@@ -74,5 +74,5 @@ class Admin::ProfilesController < Admin::BaseController
   def get_profile
     @profile = Profile.from_param params[:id]
   end
- 
+
 end
