@@ -35,8 +35,8 @@ class ApplicationController < ActionController::Base
     @themes     = Theme.enabled.includes(:seo, :activities).order(:position)
     @main_pages = MainPage.no_parent.enabled.includes(:seo).order(:position)
     @profiles   = Profile.enabled.order(:position)
-    @key_number_page  = MainPage.enabled.find_by(key: 'key_numbers')
-    @membership_page  = MainPage.enabled.find_by(key: 'membership')
+    @key_number_page  = MainPage.find_by(key: 'key_number')
+    @membership_page  = MainPage.find_by(key: 'membership')
     @ess_map_page     = MainPage.find_by(key: 'ess_map')
     @contact_page     = BasicPage.find_by(key: 'contact')
     @data_policy_page = BasicPage.find_by(key: 'data_policy')
