@@ -5,7 +5,7 @@ class Setting < RailsSettings::Base
   has_one_attached :logo_primary
 
   HIGHLIGHTED_FEATURES = %w[formations posts]
-  ENABLABLE_FEATURES = %w[formations key_numbers]
+  ENABLABLE_FEATURES = %w[formations newsletter]
 
   field :pole_name, type: :string
   field :pole_address, type: :string
@@ -27,6 +27,7 @@ class Setting < RailsSettings::Base
   field :facebook, type: :string
   field :instagram, type: :string
   field :linkedin, type: :string
+  field :twitter, type: :string
   field :highlighted_feature, default: 'formations', validates: { inclusion: { in: HIGHLIGHTED_FEATURES + [''] } }, option_values: HIGHLIGHTED_FEATURES
   field :enabled_features, type: :array, default: ENABLABLE_FEATURES
   field :people_columns_count, type: :integer, default: 2
